# admin/docs/swagger.yaml
basePath: /
consumes:
- application/json
definitions:
...
  baseModel.IdReq:
    properties:
      id:
        description: 数据ID
        example: 1
        type: integer
    type: object
  baseModel.PageRes:
    properties:
      list:
        description: 分页数据
      total:
        description: 数据总量
        example: 100
        type: integer
    type: object
  baseModel.Req:
    type: object
  baseModel.ResBody:
    properties:
      code:
        description: 响应码
        example: S000/F000
        type: string
      data:
        description: 响应数据
      msg:
        description: 响应消息
        example: Success/Fail
        type: string
      unPop:
        description: 免弹窗提示
        example: true
        type: boolean
    type: object
...
  baseModel.SelectRes:
    properties:
      label:
        description: 展示名
        example: Name
        type: string
      value:
        description: 展示值
        example: Value
        type: string
    type: object
  baseModel.SortReq:
    properties:
      id:
        description: 数据ID
        example: 1
        type: integer
      sort:
        description: 序号
        example: 1
        type: integer
    required:
    - id
    type: object
  baseModel.SortRes:
    properties:
      id:
        description: 默认数据ID
        example: 1
        type: integer
      name:
        description: 权限名称，界面展示，建议与界面导航一致
        example: Account
        type: string
      sort:
        description: 权限排序
        example: 1
        type: integer
    type: object
  baseModel.Tree:
    properties:
      children:
        description: 子树
        items:
          $ref: '#/definitions/baseModel.Tree'
        type: array
      expand:
        description: 拓展信息
      id:
        description: 表示树数据ID
        example: 1
        type: integer
      key:
        description: 树键
        example: ROOT
        type: string
      level:
        description: 表示树等级
        example: "1"
        type: string
      title:
        description: 树标题
        example: ROOT
        type: string
    type: object
...
  platModel.DictAddReq:
    properties:
      choose:
        description: 是否可被选择 0可选择 1不可选择
        enum:
        - "0"
        - "1"
        example: "0"
        type: string
      groupKey:
        description: 字典分组KEY
        example: serviceCode
        maxLength: 32
        type: string
      label:
        description: 字段名称
        example: PlatBase
        maxLength: 64
        type: string
      labelEn:
        description: 字段名称（英文）
        example: PlatBase
        maxLength: 64
        type: string
      remark:
        description: 字典描述
        example: Business type dictionary
        maxLength: 128
        type: string
      val:
        description: 字典值（字符型）
        example: "1"
        maxLength: 32
        type: string
    required:
    - choose
    - groupKey
    - label
    - labelEn
    - val
    type: object
  platModel.DictBroReq:
    properties:
      groupKey:
        description: 字典分组KEY
        example: serviceCode
        type: string
    required:
    - groupKey
    type: object
  platModel.DictEditReq:
    properties:
      choose:
        description: 是否可被选择 0可选择 1不可选择
        enum:
        - "0"
        - "1"
        example: "0"
        type: string
      id:
        description: 数据ID
        example: 1
        type: integer
      label:
        description: 字段名称
        example: PlatBase
        maxLength: 64
        type: string
      labelEn:
        description: 字段名称（英文）
        example: PlatBase
        maxLength: 64
        type: string
      remark:
        description: 字典描述
        example: Business type dictionary
        maxLength: 128
        type: string
    required:
    - choose
    - id
    - label
    - labelEn
    type: object
  platModel.DictGetRes:
    properties:
      choose:
        description: 是否可被选择 0可选择 1不可选择
        example: "0"
        type: string
      groupKey:
        description: 字典分组KEY
        example: serviceCode
        type: string
      id:
        description: 数据ID
        example: 1
        type: integer
      label:
        description: 字段名称
        example: PlatBase
        type: string
      labelEn:
        description: 字段名称（英文）
        example: PlatBase
        type: string
      remark:
        description: 字典描述
        example: Remark
        type: string
      val:
        description: 字典值（字符型）
        example: "1"
        type: string
    type: object
  platModel.DictNextValReq:
    properties:
      groupKey:
        description: 字典分组KEY
        example: serviceCode
        type: string
    required:
    - groupKey
    type: object
  platModel.DictPageReq:
    properties:
      choose:
        description: 字典是否可选择
        example: "1"
        type: string
      current:
        description: 当前页
        example: 1
        type: integer
      groupKey:
        description: 需要查询的字典分组
        example: serviceCode
        type: string
      pageSize:
        description: 单页数量
        example: 10
        type: integer
    type: object
  platModel.DictPageRes:
    properties:
      choose:
        description: 是否可被选择 0可选择 1不可选择
        example: "0"
        type: string
      groupKey:
        description: 字典分组KEY
        example: serviceCode
        type: string
      id:
        description: 数据ID
        example: 1
        type: integer
      label:
        description: 字段名称
        example: PlatBase
        type: string
      labelEn:
        description: 字段名称（英文）
        example: PlatBase
        type: string
      mark:
        description: 变更标识 0可变更 1禁止变更
        example: "1"
        type: string
      remark:
        description: 字典描述
        example: Remark
        type: string
      sort:
        description: 字典排序
        example: 0
        type: integer
      val:
        description: 字典值（字符型）
        example: "1"
        type: string
    type: object
  platModel.DictReadReq:
    properties:
      groupKeys:
        description: 需要查询的字典分组
        example:
        - serviceCode
        - responseType
        items:
          type: string
        type: array
    required:
    - groupKeys
    type: object
  platModel.DictReadRes:
    properties:
      list:
        additionalProperties:
          items:
            $ref: '#/definitions/baseModel.SelectRes'
          type: array
        description: 字典下拉列表 {'serviceCode':"[{'label':'基础','value':'1'}]"}
        type: object
      map:
        additionalProperties:
          additionalProperties:
            type: string
          type: object
        description: 字典翻译Map {'serviceCode':{'1':'基础'}}
        type: object
    type: object
  platModel.PermissionAddReq:
    properties:
      alias:
        description: 权限别名，英文，规范如下：sys，sysAccount sysAccountAdd
        example: Center
        maxLength: 32
        type: string
      level:
        description: 权限等级 1分组（一级导航）2模块（页面）3功能（按钮）第四级路由不在本表中体现
        example: "1"
        maxLength: 1
        type: string
      name:
        description: 权限名称，界面展示，建议与界面导航一致
        example: Center
        maxLength: 32
        type: string
      pid:
        description: 父级ID，默认为1
        example: 1
        type: integer
      routerIds:
        description: 路由集，提交路由ID数组
        example:
        - 1
        - 2
        - 3
        items:
          type: integer
        type: array
        uniqueItems: true
      static:
        description: 默认启用权限，0 启用 1 不启，启用后，该权限默认被分配，不可去勾
        enum:
        - "0"
        - "1"
        example: "1"
        type: string
    required:
    - alias
    - level
    - name
    - pid
    - static
    type: object
  platModel.PermissionEditReq:
    properties:
      alias:
        description: 权限别名，英文，规范如下：sys，sysAccount sysAccountAdd
        example: Center
        maxLength: 32
        type: string
      id:
        description: 数据ID
        example: 1
        type: integer
      name:
        description: 权限名称，界面展示，建议与界面导航一致
        example: Center
        maxLength: 32
        type: string
      routerIds:
        description: 路由集，提交路由ID数组
        example:
        - 1
        - 2
        - 3
        items:
          type: integer
        type: array
        uniqueItems: true
    required:
    - alias
    - id
    - name
    type: object
  platModel.PermissionGetRes:
    properties:
      alias:
        description: 权限别名，英文，规范如下：sys，sysAccount sysAccountAdd
        example: Center
        type: string
      id:
        description: 数据ID
        example: 1
        type: integer
      level:
        description: 权限等级 1分组（一级导航）2模块（页面）3功能（按钮）第四级路由不在本表中体现
        example: "1"
        type: string
      name:
        description: 权限名称，界面展示，建议与界面导航一致
        example: Center
        type: string
      pid:
        description: 父级ID，默认为1
        example: 1
        type: integer
      routerIds:
        description: 路由集，提交路由ID数组
        example:
        - 1
        - 2
        - 3
        items:
          type: integer
        type: array
      routers:
        description: 路由列表
        items:
          $ref: '#/definitions/platModel.RouterPageRes'
        type: array
      static:
        description: 默认启用权限，0 启用 1 不启，启用后，该权限默认被分配，不可去勾
        example: "1"
        type: string
    type: object
...
  platModel.RouterPageRes:
    properties:
      id:
        description: 数据ID
        example: 1
        type: integer
      logInDb:
        description: 日志入库 0 启用 1 默认不启用
        example: "0"
        type: string
      mark:
        description: 变更标识 0可变更 1禁止变更
        example: "1"
        type: string
      name:
        description: 路由名称，用于界面展示，与权限关联
        example: Login
        type: string
      remark:
        description: 其他备注信息
        example: login
        type: string
      reqLogPrint:
        description: 请求日志打印 0 打印 1 不打印
        example: "0"
        type: string
      reqLogSecure:
        description: 请求日志脱敏字段，逗号分隔，打印时允许配置
        example: phone
        type: string
      resLogPrint:
        description: 响应日志打印 0 打印 1 不打印
        example: "0"
        type: string
      resLogSecure:
        description: 响应日志脱敏字段，逗号分隔，打印时允许配置
        example: name
        type: string
      serviceCode:
        description: 业务编码（字典），为接口分组
        example: base
        type: string
      type:
        description: 免授权路由 0 免授权 1 授权
        example: "0"
        type: string
      url:
        description: 路由地址，后端访问URL，后端不在URL中携带参数，统一Post处理内容
        example: /auth/login
        type: string
    type: object
...
host: localhost:8000
info:
  contact:
    email: dev@siteol.com
    name: smart
    url: https://smart.siteol.com
  description: "![Smart](http://localhost:8000/docs/file/logo.png)\r\n\r\n# 概述\r\n\r\n轻量易用的授权基座smart，提供一个非官方RBAC授权模型的基础开箱即用中台管理服务。\r\n\r\n支持响应码自定义国际化、高粒度权限分配、接口路由配置和实时生效，同时支持部门数据权限横向访问限制。\r\n\r\n本系统的全部接口采用【POST】【application/json】方式传输数据。\r\n\r\n除开放接口以外的其他接口均需要通过【ApiKeyAuth:请求头[Token]】完成鉴权。\r\n\r\n系统技术栈：Golang、VueNext、MySQL、Redis、Gin、ArcoDesign\r\n\r\n#
    接口文档\r\n\r\n轻度依赖集成，通过静态HTML加载yaml文件进行打开接口文档。\r\n\r\n该接口文档提供Swagger[支持调试]和ReDoc[阅读增强]两个版本。\r\n\r\n[Swagger[支持调试]：/docs/swagger/index.html](http://localhost:8000/docs/swagger/index.html)
    丨 [ReDoc[阅读增强]：/docs/redoc/index.html](http://localhost:8000/docs/redoc/index.html)\r\n\r\n##
    生成说明\r\n\r\n- 首次集成\r\n\r\ngo install github.com/swaggo/swag/cmd/swag@latest\r\n\r\n-
    格式化注释代码 (选用)\r\n\r\nswag fmt\r\n\r\n- 初始化swagger.yaml文件\r\n\r\nswag init\r\n\r\n-
    如果像本工程一样依赖了api.md和独立的api.go文件描述项目（推荐，后续持续使用）\r\n\r\nswag init -g api.go --md .\r\n\r\n-
    删除多余的生成\r\n\r\nrm .\\docs\\docs.go\r\n\r\n删除docs.go是因为本项目中并未采用下述依赖来集成。\r\n\r\n-
    github.com/swaggo/swag\r\n- github.com/swaggo/gin-swagger\r\n- github.com/swaggo/files\r\n\r\n而是通过HTML模板+JS+YAML引入集成，移除docs.go用于避免依赖编译报错。\r\n\r\n源帮助页：[https://github.com/swaggo/swag/blob/master/README_zh-CN.md](https://github.com/swaggo/swag/blob/master/README_zh-CN.md)\r\n"
  title: smart
  version: "1.0"
  x-logo:
    altText: smart
...
paths:
...
  /plat/dict/add:
    post:
      consumes:
      - application/json
      description: 在字典分组下新建字典
      operationId: AddDict 字典新建
      parameters:
      - description: 请求
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/platModel.DictAddReq'
      produces:
      - application/json
      responses:
        "200":
          description: 响应成功
          schema:
            allOf:
            - $ref: '#/definitions/baseModel.ResBody'
            - properties:
                data:
                  type: boolean
              type: object
      security:
      - Token: []
      summary: 字典新建
      tags:
      - 数据字典
  /plat/dict/bro:
    post:
      consumes:
      - application/json
      description: 获取字典排序数据
      operationId: BroDict 字典排序数据
      parameters:
      - description: 请求
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/platModel.DictBroReq'
      produces:
      - application/json
      responses:
        "200":
          description: 响应成功
          schema:
            allOf:
            - $ref: '#/definitions/baseModel.ResBody'
            - properties:
                data:
                  $ref: '#/definitions/baseModel.SortRes'
              type: object
      security:
      - Token: []
      summary: 排序数据
      tags:
      - 数据字典
  /plat/dict/del:
    post:
      consumes:
      - application/json
      description: 字典封存处理
      operationId: DelDict 字典封存
      parameters:
      - description: 请求
        in: body
        name: req
        required: true
        schema:
          items:
            $ref: '#/definitions/baseModel.IdReq'
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: 响应成功
          schema:
            allOf:
            - $ref: '#/definitions/baseModel.ResBody'
            - properties:
                data:
                  type: boolean
              type: object
      security:
      - Token: []
      summary: 字典封存
      tags:
      - 数据字典
  /plat/dict/edit:
    post:
      consumes:
      - application/json
      description: 在字典分组下编辑字典
      operationId: EditDict 字典编辑
      parameters:
      - description: 请求
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/platModel.DictEditReq'
      produces:
      - application/json
      responses:
        "200":
          description: 响应成功
          schema:
            allOf:
            - $ref: '#/definitions/baseModel.ResBody'
            - properties:
                data:
                  type: boolean
              type: object
      security:
      - Token: []
      summary: 字典编辑
      tags:
      - 数据字典
  /plat/dict/get:
    post:
      consumes:
      - application/json
      description: 查询字典详情
      operationId: GetDict 字典详情
      parameters:
      - description: 请求
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/baseModel.IdReq'
      produces:
      - application/json
      responses:
        "200":
          description: 响应成功
          schema:
            allOf:
            - $ref: '#/definitions/baseModel.ResBody'
            - properties:
                data:
                  $ref: '#/definitions/platModel.DictGetRes'
              type: object
      security:
      - Token: []
      summary: 字典详情
      tags:
      - 数据字典
  /plat/dict/nextVal:
    post:
      consumes:
      - application/json
      description: 在字典分组下读取下一个Val的建议值
      operationId: NextDictVal 字典NextVal建议
      parameters:
      - description: 请求
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/platModel.DictNextValReq'
      produces:
      - application/json
      responses:
        "200":
          description: 响应成功
          schema:
            allOf:
            - $ref: '#/definitions/baseModel.ResBody'
            - properties:
                data:
                  type: integer
              type: object
      security:
      - Token: []
      summary: 字典NextVal建议
      tags:
      - 数据字典
  /plat/dict/page:
    post:
      consumes:
      - application/json
      description: 查询字典分页数据
      operationId: PageDict 字典分页
      parameters:
      - description: 请求
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/platModel.DictPageReq'
      produces:
      - application/json
      responses:
        "200":
          description: 响应成功
          schema:
            allOf:
            - $ref: '#/definitions/baseModel.ResBody'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/baseModel.PageRes'
                  - properties:
                      list:
                        items:
                          $ref: '#/definitions/platModel.DictPageRes'
                        type: array
                    type: object
              type: object
      security:
      - Token: []
      summary: 字典分页
      tags:
      - 数据字典
  /plat/dict/read:
    post:
      consumes:
      - application/json
      description: 获取字典下拉列表以及关联键值Map
      operationId: ReadDict 读取字典
      parameters:
      - description: 语言，不传默认为zh-CN
        in: header
        name: Lang
        type: string
      - description: 请求
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/platModel.DictReadReq'
      produces:
      - application/json
      responses:
        "200":
          description: 响应成功
          schema:
            allOf:
            - $ref: '#/definitions/baseModel.ResBody'
            - properties:
                data:
                  $ref: '#/definitions/platModel.DictReadRes'
              type: object
      security:
      - Token: []
      summary: 读取字典
      tags:
      - 数据字典
  /plat/dict/sort:
    post:
      consumes:
      - application/json
      description: 字典分组下字典排序处理
      operationId: SortDict 字典排序处理
      parameters:
      - description: 请求
        in: body
        name: req
        required: true
        schema:
          items:
            $ref: '#/definitions/baseModel.SortReq'
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: 响应成功
          schema:
            allOf:
            - $ref: '#/definitions/baseModel.ResBody'
            - properties:
                data:
                  type: boolean
              type: object
      security:
      - Token: []
      summary: 排序处理
      tags:
      - 数据字典
  /plat/permission/add:
    post:
      consumes:
      - application/json
      description: 新建权限配置
      operationId: AddPermission 权限新建
      parameters:
      - description: 请求
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/platModel.PermissionAddReq'
      produces:
      - application/json
      responses:
        "200":
          description: 响应成功
          schema:
            allOf:
            - $ref: '#/definitions/baseModel.ResBody'
            - properties:
                data:
                  type: boolean
              type: object
      security:
      - Token: []
      summary: 权限新建
      tags:
      - 访问权限
  /plat/permission/bro:
    post:
      consumes:
      - application/json
      description: 同级权限列表
      operationId: BroPermission 同级权限
      parameters:
      - description: 语言，不传默认为zh-CN
        in: header
        name: Lang
        type: string
      - description: 请求
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/baseModel.IdReq'
      produces:
      - application/json
      responses:
        "200":
          description: 响应成功
          schema:
            allOf:
            - $ref: '#/definitions/baseModel.ResBody'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/baseModel.SortRes'
                  type: array
              type: object
      security:
      - Token: []
      summary: 同级权限
      tags:
      - 访问权限
  /plat/permission/del:
    post:
      consumes:
      - application/json
      description: 权限封存处理
      operationId: DelPermission 权限封存
      parameters:
      - description: 语言，不传默认为zh-CN
        in: header
        name: Lang
        type: string
      - description: 请求
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/baseModel.IdReq'
      produces:
      - application/json
      responses:
        "200":
          description: 响应成功
          schema:
            allOf:
            - $ref: '#/definitions/baseModel.ResBody'
            - properties:
                data:
                  type: boolean
              type: object
      security:
      - Token: []
      summary: 权限封存
      tags:
      - 访问权限
  /plat/permission/edit:
    post:
      consumes:
      - application/json
      description: 在权限分组下编辑权限
      operationId: EditPermission 权限编辑
      parameters:
      - description: 请求
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/platModel.PermissionEditReq'
      produces:
      - application/json
      responses:
        "200":
          description: 响应成功
          schema:
            allOf:
            - $ref: '#/definitions/baseModel.ResBody'
            - properties:
                data:
                  type: boolean
              type: object
      security:
      - Token: []
      summary: 权限编辑
      tags:
      - 访问权限
  /plat/permission/get:
    post:
      consumes:
      - application/json
      description: 查询权限详情
      operationId: GetPermission 权限详情
      parameters:
      - description: 请求
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/baseModel.IdReq'
      produces:
      - application/json
      responses:
        "200":
          description: 响应成功
          schema:
            allOf:
            - $ref: '#/definitions/baseModel.ResBody'
            - properties:
                data:
                  $ref: '#/definitions/platModel.PermissionGetRes'
              type: object
      security:
      - Token: []
      summary: 权限详情
      tags:
      - 访问权限
  /plat/permission/sort:
    post:
      consumes:
      - application/json
      description: 同级权限排序功能
      operationId: SortPermission权限排序
      parameters:
      - description: 请求
        in: body
        name: req
        required: true
        schema:
          items:
            $ref: '#/definitions/baseModel.SortReq'
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: 响应成功
          schema:
            allOf:
            - $ref: '#/definitions/baseModel.ResBody'
            - properties:
                data:
                  type: boolean
              type: object
      security:
      - Token: []
      summary: 权限排序
      tags:
      - 访问权限
  /plat/permission/tree:
    post:
      consumes:
      - application/json
      description: 查询权限树数据
      operationId: PagePermission 权限树
      parameters:
      - description: 请求
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/baseModel.Req'
      produces:
      - application/json
      responses:
        "200":
          description: 响应成功
          schema:
            allOf:
            - $ref: '#/definitions/baseModel.ResBody'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/baseModel.Tree'
                  type: array
              type: object
      security:
      - Token: []
      summary: 权限树
      tags:
      - 访问权限
...
securityDefinitions:
  Token:
    in: header
    name: Token
    type: apiKey
swagger: "2.0"
...